Add unpack feature
