+++ docs/updating.md
@@ -2,16 +2,6 @@
 
 If you want to improve or update an existing recipe, please follow this guide.
 
-Please **do not** simply change files inside `uncompressed` - Ferdi only uses the packaged archives, the uncompressed files are only to improve accessibility of the repository.
-
-## 0. Find the recipe ID of the recipe you want to update
-
-Before you can start updating the recipe, you'll need to know its internal ID. The ID is a string of lowercase letters and hyphens that Ferdi uses to store and manage them.
-
-To find the recipe ID, you can search <https://github.com/getferdi/recipes/blob/master/all.json> for the recipe you want to edit.
-
-In the rest of this guide, always replace "&lt;recipe id>" with the ID of the recipe you want to edit.
-
 ## 1. Implement your updates
 
 We highly suggest that you implement and test your updates to the recipe inside Ferdi before implementing them into the recipe repository. This will save you time on debugging and repackaging the recipe later.
@@ -21,32 +11,22 @@ You can find the recipe source Ferdi uses at:
   * Windows: `%appdata%/Ferdi/recipes/<recipe id>/`
   * Linux: `~/.config/Ferdi/recipes/<recipe id>`
 
+> If you can't find the recipe's folder, please make sure that you have created a service inside Ferdi with that recipe - otherwise Ferdi will not load it into the folder
+
 After changing the files in that folder, you can reload the service inside Ferdi using `CTRL/CMD + R`.
 
 ## 2. Setting up the recipe repository
 
-- Fork https://github.com/getferdi/recipes and clone it to your computer
-- Open a terminal in the `scripts/` folder of that repository
-- Run `npm install` to install all dependencies
+Simply fork https://github.com/getferdi/recipes. You may also want to clone the repository to your computer to edit it locally.
 
-This will setup the repository so that you can work on updating it.
+## 3. Updating the recipe inside the repository
 
-## 3. Loading and updating the recipe
-
-Again, please **do not** edit any of the files located inside the `uncompressed` folder! Before editing, you'll need to "load" the recipe you want to edit using
-```JavaScript
-npm run load <recipe id>
-```
-(e.g. `npm run load gmail`)
-
-Your recipe is now loaded into `scripts/recipe_src`. You can now make changes to the files **in that folder*.
+Simply add your changes to the recipe inside this repository.
 
 ## 4. Updating the version number
 
-In order for Ferdi to update its recipe, you'll need to bump its version number. Increase the `version` inside `scripts/recipe_src/package.json`.
+This might be the __most important step__ of this guide. You'll need to increase the version number inside recipe's `package.json`.
 
-## 5. Re-packaging the recipe
+Ferdi uses this version to determine if it should update the local copy of the recipe. If you do not increase the version number, no Ferdi user will get your updated recipe 😔
 
-You can now run `npm run package` to repackage your updated files. Your changes will be automatically copied into the recipe source at `uncompressed` and compressed into `archives/<recipe id>.tar.gz` - no need to update anything manually!
-
-## 6. Commit your changes and create a PR to <https://github.com/getferdi/recipes>
+## 5. Commit your changes and create a PR to <https://github.com/getferdi/recipes>
