--- scripts/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "ferdi-recipes-scripts",
-  "version": "1.0.0",
-  "description": "Scripts for managing the ferdi-recipes repository",
-  "main": "index.js",
-  "repository": "https://github.com/getferdi/recipes",
-  "author": "The Ferdi Team",
-  "license": "MIT",
-  "dependencies": {
-    "dir-compare": "^1.7.3",
-    "fs-extra": "^8.1.0",
-    "git-url-parse": "^11.1.2",
-    "image-size": "^0.8.3",
-    "node-fetch": "^2.6.0",
-    "open": "^7.0.3",
-    "semver": "^6.3.0",
-    "targz": "^1.0.1"
-  },
-  "scripts": {
-    "package": "node package.js",
-    "verify": "node verify.js",
-    "verify-all": "node verify-all.js",
-    "update": "node update.js",
-    "github": "node add_github.js",
-    "gh-load": "node gh_load.js",
-    "load": "node load.js",
-    "create": "node create.js"
-  }
-}
